Feat: Flow Order product

+++ src/main/java/iuh/fit/se/ace_store/dto/request/ProductSearchRequest.java
@@ -2,12 +2,14 @@ package iuh.fit.se.ace_store.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductSearchRequest {
     private String keyword;
     private String category;
-    private Double minPrice;
-    private Double maxPrice;
+    private BigDecimal minPrice;
+    private BigDecimal maxPrice;
     private Integer page;
     private Integer size;
     private String sortBy;
